+++ History.md
@@ -1,5 +1,6 @@
 ### vNext ###
 - Added suppression for member declarations in XAML files (fixes issue #6)
+- Added new setting to allow to ignore overriding type members (implements #3)
 
 ### 1.4.0 ###
 - ReSharper 2016.1 support